Move BufferedActions struct to own file ADIOS2File.hpp
